Remove warnings
only possible for HTML by turning off the HTML validation.
This is not an HTML project so it does
no harm and I'm buggered if I'm going to waste my time trying
to convert perfectly good tables to divs and csss just because
some grasshole decided that it was too easy to have simple
tables.